[refactor] Introduce BaseForwardModel
