+++ chainer_chemistry/models/prediction/__init__.py
@@ -1,5 +1,7 @@
+from chainer_chemistry.models.prediction import base  # NOQA
 from chainer_chemistry.models.prediction import classifier  # NOQA
 from chainer_chemistry.models.prediction import regressor  # NOQA
 
+from chainer_chemistry.models.prediction.base import BaseForwardModel  # NOQA
 from chainer_chemistry.models.prediction.classifier import Classifier  # NOQA
 from chainer_chemistry.models.prediction.regressor import Regressor  # NOQA
